+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 * Parse target Django version from ``pyproject.toml``.
   Now, if you don’t specify a version with ``--target-version``, django-upgrade will try to parse your minimum-supported target Django version from ``project.dependencies`` in ``pyproject.toml``.
-  It supports several common formats, like ``django >= 5.2``.
+  It supports several common formats, like ``django>=5.2,<6.0``.
 
 * Add Django 5.2+ fixer ``postgres_aggregate_order_by`` to rewrite PostgreSQL aggregate functions using the old argument name ``ordering`` to the new name ``order_by``.